Implemented wall jumping
Since we did not have an animation for wall jumping I had to make my own. It isn't great but I think it will do.  Currently the wall jump enables the player to get more and more height because gravity is so low. Once gravity is implemented we can adjust the wall jump to prevent the player from climbing the wall all the way to the top.